refactor: 식별자 참조 변경으로 인한 수정 사항 적용
