+++ src/main/java/app/bottlenote/review/repository/ReviewQueryRepository.java
@@ -1,6 +1,7 @@
 package app.bottlenote.review.repository;
 
 import app.bottlenote.review.domain.Review;
+import app.bottlenote.review.repository.custom.CustomReviewQueryRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
