+++ src/main/java/app/bottlenote/review/repository/ReviewQuerySupporter.java
@@ -19,7 +19,7 @@ import static app.bottlenote.user.domain.QUser.user;
 @Component
 public class ReviewQuerySupporter {
 
-	protected ConstructorExpression<ReviewsDetailInfo.ReviewInfo> reviewInfoConstructor(Long userId) {
+	public ConstructorExpression<ReviewsDetailInfo.ReviewInfo> reviewInfoConstructor(Long userId) {
 		return Projections.constructor(
 			ReviewsDetailInfo.ReviewInfo.class,
 			user.id.as("userId"),
@@ -41,7 +41,7 @@ public class ReviewQuerySupporter {
 		);
 	}
 
-	protected BooleanExpression hasReplyByMeSubquery(Long userId) {
+	public BooleanExpression hasReplyByMeSubquery(Long userId) {
 		return Expressions.asBoolean(
 			JPAExpressions
 				.selectOne()
@@ -51,7 +51,7 @@ public class ReviewQuerySupporter {
 		).as("hasReplyByMe");
 	}
 
-	protected BooleanExpression isLikeByMeSubquery(Long userId) {
+	public BooleanExpression isLikeByMeSubquery(Long userId) {
 		return Expressions.asBoolean(
 			JPAExpressions
 				.selectOne()
