--- src/main/java/app/bottlenote/review/repository/ReviewTastingTagRepository.java
@@ -1,9 +1,0 @@
-package app.bottlenote.review.repository;
-
-import app.bottlenote.review.domain.ReviewTastingTag;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ReviewTastingTagRepository extends CrudRepository<ReviewTastingTag, Long> {
-}
